requirements.txt
fastapi
uvicorn
pydantic[email]
PyJWT
python-decouple